fix: ensure that alerts created as disabled don't actually fire
